requirements.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
